refactor(packages): reorganize package structure
