--- src/main/java/study/lotto/automatic/domain/LottoGenerator.java
@@ -1,5 +1,0 @@
-package study.lotto.automatic.domain;
-
-public interface LottoGenerator {
-    LottoNumbers generate();
-}
