+++ src/main/java/study/lotto/domain/lottomachine/LottoGenerator.java
@@ -1,0 +1,7 @@
+package study.lotto.domain.lottomachine;
+
+import study.lotto.domain.LottoNumbers;
+
+public interface LottoGenerator {
+    LottoNumbers generate();
+}
