+++ src/main/java/study/lotto/domain/lottomachine/LottoMachine.java
@@ -1,4 +1,4 @@
-package study.lotto.automatic.domain;
+package study.lotto.domain.lottomachine;
 
 import java.math.BigDecimal;
 import java.util.ArrayList;
@@ -6,6 +6,7 @@ import java.util.List;
 import java.util.Objects;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
+import study.lotto.domain.Lotto;
 
 public class LottoMachine {
     private final LottoGenerator lottoGenerator;
