+++ src/main/java/study/lotto/domain/sorter/IntegerSorter.java
@@ -1,4 +1,4 @@
-package study.lotto.automatic.domain.sorter;
+package study.lotto.domain.sorter;
 
 import java.util.List;
