+++ src/test/java/study/lotto/domain/sorter/IntegerAscendingSorterTest.java
@@ -1,4 +1,4 @@
-package study.lotto.automatic.domain.sorter;
+package study.lotto.domain.sorter;
 
 import static org.assertj.core.api.Assertions.assertThat;
